refactor(agent): extract chat submit enrichment
